+++ rich/table.py
@@ -498,7 +498,13 @@ class Table(JupyterMixin):
             text: TextType, style: StyleType, justify: "JustifyMethod" = "center"
         ) -> "RenderResult":
             render_text = (
-                console.render_str(text, style=style, highlight=False)
+                console.render_str(
+                    text,
+                    style=style,
+                    highlight=False,
+                    markup=render_options.markup,
+                    emoji=render_options.emoji,
+                )
                 if isinstance(text, str)
                 else text
             )